fix: use golang in workspace_read
