+++ file-summarizer/tools/load_text.py
